+++ src/main/scala/firrtl/passes/memlib/MemIR.scala
@@ -11,7 +11,7 @@ case class DefAnnotatedMemory(
     info: Info,
     name: String,
     dataType: Type,
-    depth: Int,
+    depth: BigInt,
     writeLatency: Int,
     readLatency: Int,
     readers: Seq[String],
